Renamed Users file, edit mock database and fix Delete Button In following commits, we must not submit mock database again Related to issue: #7
